refactor: [EL-4839] refactor the code according comments
